Rename twitter_activity -> twitter and twitter -> twitter_legacy
