--- temba/channels/types/twitter_activity/__init__.py
@@ -1,1 +1,0 @@
-from .type import TwitterActivityType  # noqa
