+++ temba/channels/types/twitter_legacy/__init__.py
@@ -1,0 +1,1 @@
+from .type import TwitterLegacyType  # noqa
